sprzatanie + potencjalnie liczenie atrakcji w bufferze
