+++ graphGenerator.py
@@ -1,67 +1,25 @@
 import arcpy
+from HelperClasses import Node, Edge, Graph
+import time
 
-class Edge:
-    def __init__(self, id, id_from, id_to, length):
-        self.id = id
-        self.id_from = id_from
-        self.id_to = id_to
-        self.length = length
+def prepare_data(kopia_drogi_torun, atrakcje):
+    """
+    1. Tworzymy buffera dla dróg
+    2. Spatial join łączy warstwy jeżeli punkty mają styczność z bufferami, 
+    przy okazji generuje pole Join_Count z liczbą punktow w bufferze
+    3. Łączymy warstwy po ID1 i dodajemy join count do drog (warto działać na kopii warstwy)
+    """
+    bufferedDrogi = arcpy.analysis.Buffer(kopia_drogi_torun, "bufferedDrogi.shp", "25 meters", dissolve_option="NONE", dissolve_field=None, method="PLANAR")
+    joined = arcpy.analysis.SpatialJoin(bufferedDrogi, atrakcje, "attractions_near_roads", "JOIN_ONE_TO_ONE", "KEEP_ALL", None, "INTERSECT")
+    arcpy.management.JoinField(kopia_drogi_torun, "ID1", joined, "ID1", ["Join_Count"])
+    return kopia_drogi_torun
 
-    def __repr__(self):
-        return f"Edge {self.id} from {self.id_from} to {self.id_to}"
-
-    def __hash__(self):
-        return hash((self.id_from, self.id_to, self.length))
-
-    def __eq__(self, other):
-        return (self.id_from, self.id_to, self.length) == (other.id_from, other.id_to, other.length)
-
-class Node:
-    def __init__(self, x, y):
-        self.x = x
-        self.y = y
-        self.id = self.generate_id()
-        self.edges = set()
-
-    def generate_id(self):
-        return f"{self.x}_{self.y}"
-
-    def add_edge(self, edge):
-        self.edges.add(edge)
-
-    def __repr__(self):
-        return f"Node {self.id}"
-
-class Graph:
-    def __init__(self):
-        self.graph = {}
-
-    def add_node(self, node):
-        if node.id not in self.graph:
-            self.graph[node.id] = node
-
-    def add_edge(self, edge):
-        if edge.id_from in self.graph:
-            self.graph[edge.id_from].add_edge(edge)
-        if edge.id_to in self.graph:
-            reversed_edge = Edge(edge.id, edge.id_to, edge.id_from, edge.length)
-            self.graph[edge.id_to].add_edge(reversed_edge)
-
-    def __repr__(self):
-        for key, value in self.graph.items():
-            print(f"{key}: {value.edges}")
-            
-arcpy.env.workspace = r"C:\Users\filo1\Desktop\szkola_sem5\PAG2\proj1\dane\torun\drogi"
-arcpy.env.overwriteOutput = True
-
-graph = Graph()
-drogiTorun = arcpy.ListFeatureClasses("*.shp")
-print(drogiTorun)
-
-for featureClass in drogiTorun:
-    with arcpy.da.SearchCursor(featureClass, ["SHAPE@"]) as cursor:
+def generate_graph(warstwa_drog):
+    graph = Graph()
+    with arcpy.da.SearchCursor(warstwa_drog, ["SHAPE@", "Join_Count"]) as cursor:
         for i, row in enumerate(cursor):
             geometry = row[0]
+            attraction_number = row[1]
             length = geometry.length
             firstX = round(geometry.firstPoint.X)
             firstY = round(geometry.firstPoint.Y)
@@ -74,7 +32,24 @@ for featureClass in drogiTorun:
             graph.add_node(firstNode)
             graph.add_node(lastNode)
 
-            edge = Edge(i, firstNode.id, lastNode.id, length)
+            edge = Edge(i, firstNode.id, lastNode.id, length, attraction_number)
             graph.add_edge(edge)
+    return graph
 
-print(graph)
+
+if __name__ == "__main__":
+    start = time.time()
+    arcpy.env.workspace = r"C:\Users\filo1\Desktop\szkola_sem5\PAG2\proj1\arcgis_proj\MyProject.gdb"
+    arcpy.env.overwriteOutput = True
+
+    atrakcjeLayer = arcpy.ListFeatureClasses("attr*")[0]
+    drogi_torun = arcpy.ListFeatureClasses("L4*")[0]
+
+    # kopia_drogi_torun = arcpy.management.CopyFeatures(drogi_torun, "kopia_drogi_torun")
+    # kopia_drogi_torun = prepare_data(kopia_drogi_torun, atrakcjeLayer)
+
+    graph = generate_graph(drogi_torun)
+    
+    print(graph.graph)
+    end = time.time()
+    print(end - start)
